+++ src/factor_instances_provider/factor_instances_provider.rs
@@ -1,201 +1,58 @@
 #![allow(unused)]
 #![allow(unused_variables)]
 
-use std::ops::Range;
-
 use crate::prelude::*;
 
-pub struct NextDerivationBasedOnProfileIndexAnalyzer {
-    local_offsets: HashMap<UnquantifiedUnindexDerivationRequest, usize>,
-    profile_snapshot: Profile,
+/// A provider of FactorInstances, reading them from the cache if present,
+/// else if missing derives many instances in Abundance and caches the
+/// not requested ones, and returns one matching the requested ones.
+pub struct FactorInstancesProvider {
+    /// The purpose of the requested FactorInstances.
+    purpose: FactorInstancesRequestPurpose,
+
+    /// If no cache present, a new one is created and will be filled.
+    cache: Arc<PreDerivedKeysCache>,
+
+    /// If we did not find any cached keys at all, use the next index
+    /// analyser to get the "next index" - which uses the Profile
+    /// if present.
+    next_index_assigner: NextDerivationIndexAnalyzer,
+
+    /// GUI hook used for KeysCollector if we need to derive more
+    /// factor instances.
+    derivation_interactors: Arc<dyn KeysDerivationInteractors>,
 }
 
-impl NextDerivationBasedOnProfileIndexAnalyzer {
-    pub fn next(&self, unindexed_request: UnquantifiedUnindexDerivationRequest) -> HDPathValue {
-        todo!()
-    }
-}
+/// ==================
+/// *** CTOR ***
+/// ==================
 
-/// With known start index and quantity
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct QuantifiedDerivationRequestWithStartIndex {
-    pub factor_source_id: FactorSourceIDFromHash,
-    pub network_id: NetworkID,
-    pub entity_kind: CAP26EntityKind,
-    pub key_kind: CAP26KeyKind,
-    pub key_space: KeySpace,
-    pub quantity: usize,
-    pub start_base_index: HDPathValue,
-}
-impl QuantifiedDerivationRequestWithStartIndex {
-    fn new(
-        factor_source_id: FactorSourceIDFromHash,
-        network_id: NetworkID,
-        entity_kind: CAP26EntityKind,
-        key_kind: CAP26KeyKind,
-        key_space: KeySpace,
-        quantity: usize,
-        start_base_index: HDPathValue,
+impl FactorInstancesProvider {
+    pub fn new(
+        purpose: FactorInstancesRequestPurpose,
+        maybe_cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
+        maybe_profile_snapshot: impl Into<Option<Profile>>,
+        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
     ) -> Self {
-        Self {
-            factor_source_id,
-            network_id,
-            entity_kind,
-            key_kind,
-            key_space,
-            quantity,
-            start_base_index,
-        }
-    }
-}
-impl From<(QuantifiedUnindexDerivationRequest, HDPathValue)>
-    for QuantifiedDerivationRequestWithStartIndex
-{
-    fn from(value: (QuantifiedUnindexDerivationRequest, HDPathValue)) -> Self {
-        let (q, i) = value;
-        Self::new(
-            q.factor_source_id,
-            q.network_id,
-            q.entity_kind,
-            q.key_kind,
-            q.key_space,
-            q.requested_quantity(),
-            i,
-        )
-    }
-}
+        let maybe_cache = maybe_cache.into();
+        let next_index_assigner = NextDerivationIndexAnalyzer::new(maybe_profile_snapshot);
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum DeriveMore {
-    WithKnownStartIndex {
-        with_start_index: QuantifiedDerivationRequestWithStartIndex,
-        number_of_instances_needed_to_fully_satisfy_request: Option<usize>,
-    },
-    WithoutKnownLastIndex(QuantifiedUnindexDerivationRequest),
-}
-impl DeriveMore {
-    pub fn requires_profile_index_assigner(&self) -> bool {
-        match self {
-            Self::WithKnownStartIndex { .. } => false,
-            Self::WithoutKnownLastIndex(_) => true,
-        }
-    }
-    /// `None` for `WithoutKnownLastIndex`, only `Some` for `WithKnownStartIndex`
-    ///  where `if_partial_how_many_to_use_directly` is `Some`
-    pub fn number_of_instances_needed_to_fully_satisfy_request(&self) -> Option<usize> {
-        match self {
-            Self::WithKnownStartIndex {
-                number_of_instances_needed_to_fully_satisfy_request,
-                ..
-            } => *number_of_instances_needed_to_fully_satisfy_request,
-            Self::WithoutKnownLastIndex(_) => None,
-        }
-    }
-    pub fn unquantified(&self) -> UnquantifiedUnindexDerivationRequest {
-        match self {
-            Self::WithKnownStartIndex {
-                with_start_index, ..
-            } => with_start_index.clone().into(),
-            Self::WithoutKnownLastIndex(request) => request.clone().into(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct NewlyDerived {
-    key: UnquantifiedUnindexDerivationRequest,
-    /// never empty
-    to_cache: FactorInstances,
-    /// can be empty
-    pub to_use_directly: FactorInstances,
-}
-impl NewlyDerived {
-    pub fn cache_all(key: UnquantifiedUnindexDerivationRequest, to_cache: FactorInstances) -> Self {
-        Self::new(key, to_cache, FactorInstances::default())
-    }
-
-    /// # Panics if `to_cache` or to `to_use_directly` is empty.
-    pub fn some_to_use_directly(
-        key: UnquantifiedUnindexDerivationRequest,
-        to_cache: FactorInstances,
-        to_use_directly: FactorInstances,
-    ) -> Self {
-        assert!(!to_use_directly.is_empty());
-        Self::new(key, to_cache, to_use_directly)
-    }
-    /// # Panics
-    /// Panics if `to_cache` is empty.
-    /// Also panics if any FactorInstances does not match the key.
-    fn new(
-        key: UnquantifiedUnindexDerivationRequest,
-        to_cache: FactorInstances,
-        to_use_directly: FactorInstances,
-    ) -> Self {
-        assert!(to_cache
-            .factor_instances()
-            .iter()
-            .all(|factor_instance| { factor_instance.satisfies(key.clone()) }));
-
-        assert!(to_use_directly
-            .factor_instances()
-            .iter()
-            .all(|factor_instance| { factor_instance.satisfies(key.clone()) }));
+        let cache = maybe_cache.unwrap_or_else(|| Arc::new(PreDerivedKeysCache::default()));
 
         Self {
-            key,
-            to_cache,
-            to_use_directly,
+            purpose,
+            cache,
+            next_index_assigner,
+            derivation_interactors,
         }
-    }
-    pub fn key_value_for_cache(&self) -> (UnquantifiedUnindexDerivationRequest, FactorInstances) {
-        (self.key.clone(), self.to_cache.clone())
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct DerivationRequestWithRange {
-    pub factor_source_id: FactorSourceIDFromHash,
-    pub network_id: NetworkID,
-    pub entity_kind: CAP26EntityKind,
-    pub key_kind: CAP26KeyKind,
-    pub key_space: KeySpace,
-    pub range: Range<HDPathValue>,
-}
-impl HDPathComponent {
-    pub fn with_base_index_in_keyspace(base_index: u32, key_space: KeySpace) -> Self {
-        match key_space {
-            KeySpace::Securified => Self::securifying_base_index(base_index),
-            KeySpace::Unsecurified => Self::unsecurified_hardening_base_index(base_index),
-        }
-    }
-}
-impl DerivationRequestWithRange {
-    pub fn derivation_paths(&self) -> IndexSet<DerivationPath> {
-        let mut paths = IndexSet::<DerivationPath>::new();
-        for i in self.range.clone() {
-            paths.insert(DerivationPath::new(
-                self.network_id,
-                self.entity_kind,
-                self.key_kind,
-                HDPathComponent::with_base_index_in_keyspace(i, self.key_space),
-            ));
-        }
-        paths
     }
 }
 
 /// ==================
-/// *** Public API ***
+/// *** Private API ***
 /// ==================
 impl FactorInstancesProvider {
     async fn derive_more(&self, requests: IndexSet<DeriveMore>) -> Result<IndexSet<NewlyDerived>> {
-        if requests.iter().any(|x| x.requires_profile_index_assigner())
-            && self
-                .next_derivation_based_on_profile_index_analyzer
-                .is_none()
-        {
-            return Err(CommonError::ProfileIndexAssignerNotPresent);
-        }
         let with_proto_ranges = requests
             .clone()
             .into_iter()
@@ -204,11 +61,7 @@ impl FactorInstancesProvider {
                     with_start_index, ..
                 } => with_start_index,
                 DeriveMore::WithoutKnownLastIndex(ref partial) => {
-                    let next_index_assigner = self
-                        .next_derivation_based_on_profile_index_analyzer
-                        .as_ref()
-                        .expect("should have been checked before");
-                    let next = next_index_assigner.next(partial.clone().into());
+                    let next = self.next_index_assigner.next(partial.clone().into());
                     QuantifiedDerivationRequestWithStartIndex::from((partial.clone(), next))
                 }
             })
@@ -274,7 +127,12 @@ impl FactorInstancesProvider {
 
         Ok(out)
     }
+}
 
+/// ==================
+/// *** Public API ***
+/// ==================
+impl FactorInstancesProvider {
     /// Does not return ALL derived FactorInstances, but only those that are
     /// related to the purpose of the request.
     ///
@@ -353,148 +211,6 @@ impl FactorInstancesProvider {
         }
 
         Ok(FactorInstances::from(factor_instances_to_use_directly))
-    }
-}
-
-impl NextDerivationBasedOnProfileIndexAnalyzer {
-    pub fn new(profile_snapshot: Profile) -> Self {
-        Self {
-            profile_snapshot,
-            local_offsets: HashMap::new(),
-        }
-    }
-}
-pub struct FactorInstancesProvider {
-    purpose: FactorInstancesRequestPurpose,
-
-    /// If no cache present, a new one is created and will be filled.
-    cache: Arc<PreDerivedKeysCache>,
-    next_derivation_based_on_profile_index_analyzer:
-        Option<NextDerivationBasedOnProfileIndexAnalyzer>,
-
-    /// GUI hook
-    derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-}
-
-/// ==================
-/// *** CTOR ***
-/// ==================
-
-impl FactorInstancesProvider {
-    fn new(
-        purpose: FactorInstancesRequestPurpose,
-        maybe_cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
-        maybe_profile_snapshot: impl Into<Option<Profile>>,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        let maybe_cache = maybe_cache.into();
-        let next_derivation_based_on_profile_index_analyzer = maybe_profile_snapshot
-            .into()
-            .map(NextDerivationBasedOnProfileIndexAnalyzer::new);
-
-        let cache = maybe_cache.unwrap_or_else(|| Arc::new(PreDerivedKeysCache::default()));
-
-        Self {
-            purpose,
-            cache,
-            next_derivation_based_on_profile_index_analyzer,
-            derivation_interactors,
-        }
-    }
-}
-
-/// ==================
-/// *** Purposes ***
-/// ==================
-impl FactorInstancesProvider {
-    pub fn oars(
-        factor_sources: &FactorSources,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        Self::new(
-            FactorInstancesRequestPurpose::OARS {
-                factor_sources: factor_sources.clone(),
-            },
-            None,
-            None,
-            derivation_interactors,
-        )
-    }
-
-    pub fn mars(
-        factor_source: &HDFactorSource,
-        cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
-        profile_snapshot: Profile,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        Self::new(
-            FactorInstancesRequestPurpose::MARS {
-                factor_source: factor_source.clone(),
-                network_id: profile_snapshot.current_network(),
-            },
-            cache,
-            profile_snapshot,
-            derivation_interactors,
-        )
-    }
-
-    pub fn pre_derive_instance_for_new_factor_source(
-        factor_source: &HDFactorSource,
-        cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
-        profile_snapshot: Profile,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        Self::new(
-            FactorInstancesRequestPurpose::PreDeriveInstancesForNewFactorSource {
-                factor_source: factor_source.clone(),
-            },
-            cache,
-            profile_snapshot,
-            derivation_interactors,
-        )
-    }
-
-    pub fn new_virtual_unsecurified_account(
-        network_id: NetworkID,
-        factor_source: &HDFactorSource,
-        cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
-        profile_snapshot: Profile,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        Self::new(
-            FactorInstancesRequestPurpose::NewVirtualUnsecurifiedAccount {
-                network_id,
-                factor_source: factor_source.clone(),
-            },
-            cache,
-            profile_snapshot,
-            derivation_interactors,
-        )
-    }
-
-    /// Securify unsecurified Accounts
-    ///
-    /// # Panics
-    /// Panics if `UnsecurifiedEntity` is not an account
-    /// or if it is not present in `profile_snapshot`.
-    pub fn securify_unsecurified_accounts(
-        unsecurified_accounts: UnsecurifiedAccounts,
-        matrix_of_factor_sources: MatrixOfFactorSources,
-        cache: impl Into<Option<Arc<PreDerivedKeysCache>>>,
-        profile_snapshot: Profile,
-        derivation_interactors: Arc<dyn KeysDerivationInteractors>,
-    ) -> Self {
-        assert!(profile_snapshot.contains_accounts(unsecurified_accounts.clone()));
-
-        Self::new(
-            FactorInstancesRequestPurpose::SecurifyUnsecurifiedAccounts {
-                unsecurified_accounts,
-                matrix_of_factor_sources,
-            },
-            cache,
-            profile_snapshot,
-            derivation_interactors,
-        )
     }
 }
