+++ src/factor_instances_provider/mod.rs
@@ -1,6 +1,12 @@
+mod derive_more;
 #[allow(clippy::module_inception)]
 mod factor_instances_provider;
+mod factor_instances_provider_operations;
 mod factor_instances_request_purpose;
+mod next_derivation_index_analyzer;
 
+pub(crate) use derive_more::*;
 pub use factor_instances_provider::*;
+pub use factor_instances_provider_operations::*;
 pub use factor_instances_request_purpose::*;
+pub use next_derivation_index_analyzer::*;
