+++ src/recovery_securify_cache/helpers/unindex_derivation_request.rs
@@ -12,6 +12,87 @@ pub struct UnquantifiedUnindexDerivationRequest {
     pub key_space: KeySpace,
 }
 
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct DerivationRequestWithRange {
+    pub factor_source_id: FactorSourceIDFromHash,
+    pub network_id: NetworkID,
+    pub entity_kind: CAP26EntityKind,
+    pub key_kind: CAP26KeyKind,
+    pub key_space: KeySpace,
+    pub range: Range<HDPathValue>,
+}
+impl HDPathComponent {
+    pub fn with_base_index_in_keyspace(base_index: u32, key_space: KeySpace) -> Self {
+        match key_space {
+            KeySpace::Securified => Self::securifying_base_index(base_index),
+            KeySpace::Unsecurified => Self::unsecurified_hardening_base_index(base_index),
+        }
+    }
+}
+impl DerivationRequestWithRange {
+    pub fn derivation_paths(&self) -> IndexSet<DerivationPath> {
+        let mut paths = IndexSet::<DerivationPath>::new();
+        for i in self.range.clone() {
+            paths.insert(DerivationPath::new(
+                self.network_id,
+                self.entity_kind,
+                self.key_kind,
+                HDPathComponent::with_base_index_in_keyspace(i, self.key_space),
+            ));
+        }
+        paths
+    }
+}
+
+/// With known start index and quantity
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct QuantifiedDerivationRequestWithStartIndex {
+    pub factor_source_id: FactorSourceIDFromHash,
+    pub network_id: NetworkID,
+    pub entity_kind: CAP26EntityKind,
+    pub key_kind: CAP26KeyKind,
+    pub key_space: KeySpace,
+    pub quantity: usize,
+    pub start_base_index: HDPathValue,
+}
+impl QuantifiedDerivationRequestWithStartIndex {
+    fn new(
+        factor_source_id: FactorSourceIDFromHash,
+        network_id: NetworkID,
+        entity_kind: CAP26EntityKind,
+        key_kind: CAP26KeyKind,
+        key_space: KeySpace,
+        quantity: usize,
+        start_base_index: HDPathValue,
+    ) -> Self {
+        Self {
+            factor_source_id,
+            network_id,
+            entity_kind,
+            key_kind,
+            key_space,
+            quantity,
+            start_base_index,
+        }
+    }
+}
+impl From<(QuantifiedUnindexDerivationRequest, HDPathValue)>
+    for QuantifiedDerivationRequestWithStartIndex
+{
+    fn from(value: (QuantifiedUnindexDerivationRequest, HDPathValue)) -> Self {
+        let (q, i) = value;
+        Self::new(
+            q.factor_source_id,
+            q.network_id,
+            q.entity_kind,
+            q.key_kind,
+            q.key_space,
+            q.requested_quantity(),
+            i,
+        )
+    }
+}
+
 impl From<QuantifiedDerivationRequestWithStartIndex> for UnquantifiedUnindexDerivationRequest {
     fn from(value: QuantifiedDerivationRequestWithStartIndex) -> Self {
         UnquantifiedUnindexDerivationRequest::new(
